% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_cartainternacional.R
\name{download_cartainternacional}
\alias{download_cartainternacional}
\title{Title}
\usage{
download_cartainternacional(year, volume, number, dir, info_data = FALSE)
}
\arguments{
\item{info_data}{}
}
\value{

}
\description{
Title
}
