% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_conjunturaaustral.R
\name{download_conjunturaaustral}
\alias{download_conjunturaaustral}
\title{Title}
\usage{
download_conjunturaaustral(year, volume, number, dir, info_data = FALSE)
}
\arguments{
\item{info_data}{}
}
\value{

}
\description{
Title
}
