% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_conjunturaglobal.R
\name{download_conjunturaglobal}
\alias{download_conjunturaglobal}
\title{Title}
\usage{
download_conjunturaglobal(year, volume, number, dir, info_data = FALSE)
}
\arguments{
\item{info_data}{}
}
\value{

}
\description{
Title
}
