% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_conjunturainternacional.R
\name{download_conjunturainternacional}
\alias{download_conjunturainternacional}
\title{Title}
\usage{
download_conjunturainternacional(year, volume, number, dir, info_data = FALSE)
}
\arguments{
\item{full_text}{}
}
\value{

}
\description{
Title
}
