% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_contextointernacional2.R
\name{download_contextointernacional2}
\alias{download_contextointernacional2}
\title{Title}
\usage{
download_contextointernacional2(year, volume, number, dir, info_data = FALSE)
}
\arguments{
\item{info_data}{}
}
\value{

}
\description{
Title
}
