% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_estudosinternacionais.R
\name{download_estudosinternacionais}
\alias{download_estudosinternacionais}
\title{Title}
\usage{
download_estudosinternacionais(year, volume, number, dir)
}
\arguments{
\item{dir}{}
}
\value{

}
\description{
Title
}
