% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_meridiano.R
\name{download_meridiano}
\alias{download_meridiano}
\title{Title}
\usage{
download_meridiano(year, volume, number, dir, info_data = FALSE)
}
\arguments{
\item{info_data}{}
}
\value{

}
\description{
Title
}
