% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_muralinternacional.R
\name{download_muralinternacional}
\alias{download_muralinternacional}
\title{Title}
\usage{
download_muralinternacional(year, volume, number, dir, info_data = FALSE)
}
\arguments{
\item{info_data}{}
}
\value{

}
\description{
Title
}
