% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_estudosinternacionais.R
\name{download_estudosinternacionais}
\alias{download_estudosinternacionais}
\title{Download articles from Estudos Internacionais}
\usage{
download_estudosinternacionais(year, volume, number, dir, info_data = FALSE)
}
\arguments{
\item{year}{Issue Year}

\item{volume}{Issue volume}

\item{number}{Issue Number}

\item{dir}{Directory for storing files}

\item{info_data}{Whether to return or not a tibble with info on collected pdfs}
}
\value{
Nothing. It just download files based on user input
}
\description{
Download articles from Estudos Internacionais
}
\examples{
download_estudosinternacionais(year = 2020, volume = 8, number = 1, dir = '2020-v8-n1-estudosinternacionais')
}
